+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Consumption/Rules/CapacityRequirementRule.cs
@@ -12,28 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using Energinet.DataHub.MeteringPoints.Domain.Addresses;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Consumption.Rules
 {
-    public class GeoInfoReferenceRequirementRule : IBusinessRule
+    public class CapacityRequirementRule : IBusinessRule
     {
-        public GeoInfoReferenceRequirementRule(Address address)
+        public CapacityRequirementRule(Capacity? capacity, NetSettlementGroup netSettlementGroup)
         {
-            if (address == null) throw new ArgumentNullException(nameof(address));
-            IsBroken = !IsGeoInfoReferencePresent(address);
-            ValidationError = new GeoInfoReferenceIsRequiredRuleError();
+            IsBroken = netSettlementGroup != NetSettlementGroup.Zero && capacity == null;
+            ValidationError = new CapacityIsRequiredRuleError();
         }
 
         public bool IsBroken { get; }
 
         public ValidationError ValidationError { get; }
-
-        private static bool IsGeoInfoReferencePresent(Address address)
-        {
-            return !(address.GeoInfoReference == Guid.Empty || address.GeoInfoReference.HasValue == false);
-        }
     }
 }
