+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Rules/CityIsRequiredRuleError.cs
@@ -12,18 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules
 {
-    public class ConnectionTypeIsRequiredRuleError : ValidationError
+    public class CityIsRequiredRuleError : ValidationError
     {
-        public ConnectionTypeIsRequiredRuleError(NetSettlementGroup netSettlementGroup)
-        {
-            NetSettlementGroup = netSettlementGroup;
-        }
-
-        public NetSettlementGroup NetSettlementGroup { get; }
     }
 }
