+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Rules/ConnectionTypeRequirementRule.cs
@@ -17,7 +17,7 @@ using System.Globalization;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints.Rules
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules
 {
     public class ConnectionTypeRequirementRule : IBusinessRule
     {
